Cambios 29-0-19 /1

+++ src/tema9/Interface1/TrianguloEquilatero.java
@@ -9,7 +9,7 @@ package tema9.Interface1;
 
 public class TrianguloEquilatero implements PoligonoRegular{
 	float longiLados;
-	public TrianguloEquilatero(int longiLados) {
+	public TrianguloEquilatero(float longiLados) {
 		super();
 		getLongitudLados(longiLados);
 		this.longiLados = longiLados;
